Create script.js
In this code, we first get a reference to the table element and all of its header cells using document.getElementById() and querySelectorAll(), respectively. We then use a forEach() loop to add a click event listener to each header cell.

When a header cell is clicked, the callback function will be called. The function first gets the index of the clicked header cell using Array.from() and indexOf(). It then gets all of the rows in the table body using querySelectorAll(), and converts the resulting NodeList to an array using Array.from().

Next, it sorts the rows using the sort() method, which takes a comparator function that compares the text content of the cells in the clicked column using localeCompare(). We pass an options object to localeCompare() to ensure that numeric values are sorted correctly.

Finally, the sorted rows are appended back to the table body using the append() method.

With this code, the user can now click on any column header to sort the table by that column.